+++ sqlmath.mjs
@@ -304,6 +304,26 @@ function dbExecAndReturnLastBlobAsync({
     });
 }
 
+async function dbExecAndReturnLastJsonAsync(option) {
+
+// This function will exec <sql> in <db> and return last value retrieved
+// from execution as raw text.
+
+    return JSON.parse(
+        await dbExecAndReturnLastTextAsync(option)
+    );
+}
+
+async function dbExecAndReturnLastTextAsync(option) {
+
+// This function will exec <sql> in <db> and return last value retrieved
+// from execution as raw text.
+
+    return new TextDecoder().decode(
+        await dbExecAndReturnLastBlobAsync(option)
+    );
+}
+
 async function dbExecAsync({
     bindList = [],
     db,
@@ -977,6 +997,8 @@ export {
     assertOrThrow,
     dbCloseAsync,
     dbExecAndReturnLastBlobAsync,
+    dbExecAndReturnLastJsonAsync,
+    dbExecAndReturnLastTextAsync,
     dbExecAsync,
     dbFileExportAsync,
     dbFileImportAsync,